Add new scene for game choice
Added a new scene for the main window which appears after clicking Play
button in the main pane. This scene allows user to pick the game to
play.
Added a bunch of buttons (including placeholders) and implemented a
handler for swapping back and forth between the main scene and the
game-choice scene.
Some fields' name refactoring.