Update saveProfileData and add loadProfileData
Updated the saveProfileData method in ProfileData so it validates the
list of profiles passed as a parameter before saving data.
Implemented the loadProfileData method which loads profiles data to the
current session ProfileData singleton.
The functionality of both methods is based on JDOM library. The data
save and load methods are invoked on init() and stop() in the MainApp.
The application is guaranteed to save and load data using some
placeholder list during runtime to generate XML file, or a placeholder
XML file on startup to load data into the current session (if either is
missing).
Added some situational constructors to the Profile class.
Some leftover debugging code may be present.