Fix dialog for profile editing
Fixed all issues with a dialog that gets opened when editing
("creating") a profile. The dialog now properly processes the input from
the TextField and in case of any invalid name provided (void or any
combination of "empty" string) a warning alert will be shown. The dialog
doesn't close until the process is cancelled (Cancel button) by the user
or the correct input is entered and accepted (OK button).
Some minor changes to the structure of both Profile and ProfileData.
May still require some further parameter validation (constructors etc.)
and some methods needs to be moved from the inner class to
the ProfileData.