Fix active profile selection
Fixed both UI and data aspect of selecting the active profile. The app
now remembers which profile should be active, and the corresponding row
in the window for profile choice is now selected properly every time
it's brought up. The ProfileData list of profiles is updated
accordingly to the changes made in the UI.