Fix maven-shade packaging config
The graphics.jar is no longer loaded as system dependency referenced
with systemPath. It has to be installed in local maven repo now.
Resolved some issues with the path to profiles.xml (mostly for running
packaged JAR file later on).
To produce executable JAR use maven-shade-plugin (preconfigured in
pom.xml).