Implement Turbo-Button handler
Initial implementation, possibly not optimal. The turbo mode toggle
button shortens the delay between subsequent reels starting and
stopping. This approach impacts the RNG aspect of the entire process.
Delay-based synchronization of specific steps should be replaced with
the use of either proper synchronization blocks or the use of some
sentinel variables.