Change the way JARs are installed in Maven
Changed the Maven configuration and the way of adding/installing local
JARs required for this project.
- ${basedir}/lib now servers as a local Maven repo with all JARs
required as dependencies installed in it; read a README located in this
directory for more information on both installing new JARs or updating
the current repository;
- ${basedir}/JAR contains the newest versions of any local JARs that are
installed in the local Maven repo - for the reference only - the actual
JARs which are used as project dependencies are in the Maven repo;
Also updated pom.xml.