Add initial version of Info Button Dialog to Slots
Implemented a dialog (custom window) that shows when the user clicks
info button in the Slots game window.
Also added a "veil" effect to the Slots window - when the info dialog
pops up, the underlying owner window is greyed out (a Region with a
switchable visibility on top of the top level StackPane).