Update Info Button Dialog in Slots
Finalized the implementation of a dialog that pops up when clicking the
Info Button in the Slots game.
The dialog contains simplified instructions on what special symbols mean
and how nudge mechanics works. Instructions are split into two pages
that can be switched back and forth within the dialog pane.